extract `mkdir` and `mkfile` to `conftest.py`
